chore(cli): condensed all cli logic into one file
for mutations to have a better score lol
